refactor: rename from 'supabaseSync.ts' to 'syncArticles.ts'
